fixed inquirer prompts and cleaned up code

+++ server.js
@@ -1,15 +1,8 @@
 // DEPENDENCIES
 const mysql = require('mysql')
 const inquirer = require('inquirer')
-const { start } = require('repl')
 
-// VARIABLES
-let viewQuery = 'SELECT e.id as ID, e.first_name as FirstName, e.last_name as LastName, '
-viewQuery += 'role.title as Title, role.salary as Salary, department.name as Department, '
-viewQuery += 'CONCAT_WS(" ", s.first_name, s.last_name) AS Manager FROM employee e '
-viewQuery += 'LEFT JOIN employee s ON s.id = e.manager_id INNER JOIN role ON e.role_id = role.id '
-viewQuery += 'INNER JOIN department ON role.department_id = department.id'
-
+// MYSQL CONNECTION
 const connection = mysql.createConnection({
     host: 'localhost',
 
@@ -22,12 +15,21 @@ const connection = mysql.createConnection({
     database: 'employeeDB',
 });
 
+// WHEN CONNECTED TO SERVER RUN THIS
 connection.connect((err) => {
     if (err) throw err;
     console.log(`connected as id ${connection.threadId}\n`);
     startingOptions();
 });
 
+// VARIABLES
+let viewQuery = 'SELECT e.id as ID, CONCAT_WS(" ", e.first_name, e.last_name) AS Name, '
+viewQuery += 'role.title as Title, role.salary as Salary, department.name as Department, '
+viewQuery += 'CONCAT_WS(" ", s.first_name, s.last_name) AS Manager FROM employee e '
+viewQuery += 'LEFT JOIN employee s ON s.id = e.manager_id INNER JOIN role ON e.role_id = role.id '
+viewQuery += 'INNER JOIN department ON role.department_id = department.id'
+
+// FUNCTION TO START ASKING
 async function startingOptions() {
     choicesArr = ["View All Employees", "View All Employees by Department",
         "View All Employees by Manager", "Add Data", "Remove Data",
@@ -95,11 +97,11 @@ async function viewByDepartment() {
                     name: "department_id",
                     message: "What department would you like to check?: ",
                     type: "list",
-                    choices: departmentData.map(department => department.name)
+                    choices: departmentData.map(department => ({ name: department.name, value: department.id }))
                 }
             ]
         )
-        connection.query(viewQuery + ' AND department.name = ? ORDER BY e.id;', [response.department_id], (err, finalData) => {
+        connection.query(viewQuery + ' AND department.id = ? ORDER BY e.id;', [response.department_id], (err, finalData) => {
             if (err) throw err;
             console.table(finalData);
             startingOptions();
@@ -117,7 +119,7 @@ async function viewByManager() {
                     name: 'employee',
                     message: 'Choose the manager: ',
                     type: 'list',
-                    choices: employeeData.map(employee => ({name:employee.first_name + " " + employee.last_name, value:employee.id}))
+                    choices: employeeData.map(employee => ({ name: employee.first_name + " " + employee.last_name, value: employee.id }))
                 }
             ]
         )
@@ -213,38 +215,28 @@ async function addEmployee() {
                     name: "role_id",
                     message: "What role will they have?: ",
                     type: "list",
-                    choices: res.map(role => role.title)
+                    choices: roleData.map(role => ({ name: role.title, value: role.id }))
                 }
             ]
         )
-        let role_id
-        roleData.forEach((role) => {
-            if (role.title === role_response.role_id) {
-                role_id = role.id;
-            }
-        })
-
-        let manager_id
         connection.query('SELECT * FROM employee', async (err, employeeData) => {
             if (err) throw err;
-            let managerArray = employeeData.map(employee => employee.first_name + " " + employee.last_name)
+            let managerArray = employeeData.map(employee => ({ name: employee.first_name + " " + employee.last_name, value: employee.id }))
             managerArray.push("None")
             const manager_response = await inquirer.prompt(
                 [
                     {
-                        name: 'manager',
+                        name: 'manager_id',
                         message: 'Choose the manager: ',
                         type: 'list',
                         choices: managerArray
                     }
                 ]
             )
-            employeeData.forEach((employee) => {
-                if (manager_response.manager === employee.first_name + " " + employee.last_name) {
-                    manager_id = employee.id;
-                }
-            })
-            connection.query('INSERT INTO employee (first_name, last_name, role_id, manager_id) VALUES (?, ?, ?, ?);', [name_response.first_name, name_response.last_name, role_id, manager_id], (err, res) => {
+            if (manager_response.manager_id === 'None'){
+                manager_response.manager_id = null
+            }
+            connection.query('INSERT INTO employee (first_name, last_name, role_id, manager_id) VALUES (?, ?, ?, ?);', [name_response.first_name, name_response.last_name, role_response.role_id, manager_response.manager_id], (err, res) => {
                 if (err) throw err;
                 console.log('Employee Added')
                 startingOptions();
@@ -281,43 +273,31 @@ async function addOptions() {
 
 // FUNCTION TO UPDATE AN EMPLOYEE'S ROLE
 async function updateEmployeeRole() {
-    let employee_id
-    let role_id
     connection.query('SELECT * FROM employee', async (err, employeeData) => {
         if (err) throw err;
         const employee_response = await inquirer.prompt(
             [
                 {
-                    name: "employee",
+                    name: "employee_id",
                     message: "Which employee's role would you like to update?: ",
                     type: "list",
-                    choices: employeeData.map(employee => employee.first_name + " " + employee.last_name),
+                    choices: employeeData.map(employee => ({ name: employee.first_name + " " + employee.last_name, value: employee.id })),
                 }
             ]
         )
-        employeeData.forEach((employee) => {
-            if (employee_response.employee === employee.first_name + " " + employee.last_name) {
-                employee_id = employee.id;
-            }
-        })
         connection.query('SELECT * FROM role', async (err, roleData) => {
             if (err) throw err;
             const role_response = await inquirer.prompt(
                 [
                     {
-                        name: "role",
+                        name: "role_id",
                         message: "What will be the new role?: ",
                         type: "list",
-                        choices: roleData.map(role => role.title)
+                        choices: roleData.map(role => ({ name: role.title, value: role.id }))
                     }
                 ]
             )
-            roleData.forEach((role) => {
-                if (role_response.role === role.title) {
-                    role_id = role.id
-                }
-            })
-            connection.query('UPDATE employee SET role_id = ? WHERE id = ?', [role_id, employee_id], (err, res) => {
+            connection.query('UPDATE employee SET role_id = ? WHERE id = ?', [role_response.role_id, employee_response.employee_id], (err, res) => {
                 if (err) throw err;
                 console.log("Updated Employee's Role")
                 startingOptions();
@@ -328,47 +308,28 @@ async function updateEmployeeRole() {
 
 // FUNCTION TO UPDATE MANAGER
 async function updateEmployeeManager() {
-    let employee_id
-    let manager_id
     connection.query('SELECT * FROM employee', async (err, employeeData) => {
         if (err) throw err;
         const employee_response = await inquirer.prompt(
             [
                 {
-                    name: "employee",
+                    name: "employee_id",
                     message: "Which employee's manager would you like to update?: ",
                     type: "list",
-                    choices: employeeData.map(employee => employee.first_name + " " + employee.last_name),
+                    choices: employeeData.map(employee => ({ name: employee.first_name + " " + employee.last_name, value: employee.id })),
+                },
+                {
+                    name: "manager_id",
+                    message: "Who will be their new manager?: ",
+                    type: "list",
+                    choices: employeeData.map(employee => ({ name: employee.first_name + " " + employee.last_name, value: employee.id })),
                 }
             ]
         )
-        employeeData.forEach((employee) => {
-            if (employee_response.employee === employee.first_name + " " + employee.last_name) {
-                employee_id = employee.id;
-            }
-        })
-        connection.query('SELECT * FROM employee', async (err, employeeData) => {
+        connection.query('UPDATE employee SET manager_id = ? WHERE id = ?', [employee_response.manager_id, employee_response.employee_id], (err, res) => {
             if (err) throw err;
-            const employee_response = await inquirer.prompt(
-                [
-                    {
-                        name: "employee",
-                        message: "Who will be their new manager?: ",
-                        type: "list",
-                        choices: res.map(employee => employee.first_name + " " + employee.last_name),
-                    }
-                ]
-            )
-            employeeData.forEach((employee) => {
-                if (employee_response.employee === employee.first_name + " " + employee.last_name) {
-                    manager_id = employee.id;
-                }
-            })
-            connection.query('UPDATE employee SET manager_id = ? WHERE id = ?', [manager_id, employee_id], (err, res) => {
-                if (err) throw err;
-                console.log("Updated Employee's Manager")
-                startingOptions();
-            })
+            console.log("Updated Employee's Manager")
+            startingOptions();
         })
     })
 }
@@ -400,25 +361,19 @@ async function removeOptions() {
 // FUNCTION TO REMOVE EMPLOYEE
 // NOTE CANNOT DELETE MANAGERS YET
 async function removeEmployee() {
-    let employee_id
     connection.query('SELECT * FROM employee', async (err, employeeData) => {
         if (err) throw err;
         const employee_response = await inquirer.prompt(
             [
                 {
-                    name: "employee",
+                    name: "employee_id",
                     message: "Which employee is being removed?: ",
                     type: "list",
-                    choices: employeeData.map(employee => employee.first_name + " " + employee.last_name),
+                    choices: employeeData.map(employee => ({ name: employee.first_name + " " + employee.last_name, value: employee.id })),
                 }
             ]
         )
-        employeeData.forEach((employee) => {
-            if (employee_response.employee === employee.first_name + " " + employee.last_name) {
-                employee_id = employee.id;
-            }
-        })
-        connection.query('DELETE FROM employee WHERE ?', [{ "employee.id": employee_id }], (err, res) => {
+        connection.query('DELETE FROM employee WHERE ?', [{ "employee.id": employee_response.employee_id }], (err, res) => {
             if (err) throw err;
             console.log('DELETED EMPLOYEE');
             startingOptions();
@@ -428,25 +383,19 @@ async function removeEmployee() {
 
 // FUNCTION TO REMOVE ROLE
 async function removeRole() {
-    let role_id
     connection.query('SELECT * FROM role', async (err, roleData) => {
         if (err) throw err;
-        const response = await inquirer.prompt(
+        const role_response = await inquirer.prompt(
             [
                 {
-                    name: "role",
+                    name: "role_id",
                     message: "Which role is being removed?: ",
                     type: "list",
-                    choices: roleData.map(role => role.title),
+                    choices: roleData.map(role => ({name:role.title,value:role.id})),
                 }
             ]
         )
-        roleData.forEach((role) => {
-            if (response.role === role.title) {
-                role_id = role.id;
-            }
-        })
-        connection.query('DELETE FROM role WHERE ?', [{ "role.id": role_id }], (err, res) => {
+        connection.query('DELETE FROM role WHERE ?', [{ "role.id": role_response.role_id }], (err, res) => {
             if (err) throw err;
             console.log('DELETED ROLE');
             startingOptions();
@@ -456,7 +405,6 @@ async function removeRole() {
 
 // FUNCTION TO REMOVE DEPARTMENT
 async function removeDepartment() {
-    let department_id
     connection.query('SELECT * FROM department', async (err, departmentData) => {
         if (err) throw err;
         const response = await inquirer.prompt(
@@ -465,16 +413,11 @@ async function removeDepartment() {
                     name: "department",
                     message: "Which department is being removed?: ",
                     type: "list",
-                    choices: departmentData.map(department => department.name),
+                    choices: departmentData.map(department => ({name:department.name,value:department.id})),
                 }
             ]
         )
-        departmentData.forEach((department) => {
-            if (response.department === department.name) {
-                department_id = department.id;
-            }
-        })
-        connection.query('DELETE FROM department WHERE ?', [{ "department.id": department_id }], (err, res) => {
+        connection.query('DELETE FROM department WHERE ?', [{ "department.id": response.department }], (err, res) => {
             if (err) throw err;
             console.log('DELETED DEPARTMENT');
             startingOptions();
@@ -484,7 +427,6 @@ async function removeDepartment() {
 
 // FUNCTION TO GET TOTAL BUDGET OF A DEPARTMENT
 async function getBudget() {
-    let department_id
     let budget = 0
     connection.query('SELECT * FROM department', async (err, departmentData) => {
         if (err) throw err;
@@ -494,16 +436,11 @@ async function getBudget() {
                     name: "department_id",
                     message: "What department would you like to check?: ",
                     type: "list",
-                    choices: departmentData.map(department => department.name)
+                    choices: departmentData.map(department => ({name:department.name,value:department.id})),
                 }
             ]
         )
-        departmentData.forEach((department) => {
-            if (response.department_id === department.name) {
-                department_id = department.id;
-            }
-        })
-        connection.query(viewQuery + ' AND department.id =  ?', [department_id], (err, finalData) => {
+        connection.query(viewQuery + ' AND department.id =  ?', [response.department_id], (err, finalData) => {
             if (err) throw err;
             finalData.forEach((employee) => {
                 budget += employee.Salary
